pyproject.toml
[tool.poetry]
name = "turtlefun"
version = "0.1.0"
description = "Have fun with turtle graphics"
authors = ["Robert Schneider <robert@schneirob.de>"]
readme = "README.md"
license = "MIT"

[tool.poetry.dependencies]
python = "^3.12"
loguru = "^0.7.2"
click = "^8.1.7"
pillow = "^10.4.0"
prettytable = "^3.11.0"
configparser = "^7.0.0"
colorcet = "^3.1.0"
sympy = "^1.13.2"

[tool.poetry.scripts]
turtlefun = "turtlefun.turtlefun:turtlefun"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"
coverage = {extras = ["toml"], version = "^7.6.1"}
pytest-cov = "^5.0.0"
filetype = "^1.2.0"
pathlib = "^1.0.1"

[tool.coverage.paths]
source = ["src"]

[tool.coverage.run]
branch = true
source = ["turtlefun"]

[tool.coverage.report]
show_missing = true


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
